requirements.txt
#今回使用している外部ライブラリ
streamlit==1.7.0 
# click関連のエラーを消すため
# https://discuss.streamlit.io/t/attributeerror-module-click-has-no-attribute-get-os-args/23621/2
click<=8.0.4
#株価可視化アプリ用ライブラリ
yfinance==0.1.70
#azure api(物体検出)
azure-common==1.1.28
azure-cognitiveservices-vision-computervision==0.9.0
python-dotenv==0.20.0